Store window geometry and thresholds as readable settings
Window state used to live in the settings file as an opaque QByteArray, and
four detection thresholds as @Variant binary blobs - nothing a user could read,
let alone edit:

  geometry=@ByteArray(\x1\xd9\xd0\xcb\0\x3\0\0\0\0\0\0…)
  Navigation\BlackThreshold=@Variant(\0\0\0\x87?z\xe1H)

Both are plain text now: x/y/width/height/maximized under [MainWindow], and
decimals for the thresholds. Existing entries are converted on the first start
- the blob is decoded by parsing its documented layout with QDataStream rather
than feeding it to restoreGeometry() on an unshown widget, so the migration
does not depend on window-state timing. Inherited float values are rounded to
three decimals on load, because widening 0.98f to double yields
0.9800000190734863.

The QuickJump dialog wrote its size through an argument-less QSettings while
main() never set an organisation name, so that one value landed in
~/.config/Unknown Organization/TTCut-ng.conf. Setting the name fixes the cause;
the stray value is imported once and its file removed - but only when nothing
else is left in it, because that directory is shared with every other Qt
application that made the same mistake.

The value logic sits in gui/ttwindowgeometry.{h,cpp} as free functions with no
widget dependency, so tools/diag/test_window_geometry can drive all of it
headless: 51 checks covering round-trips, the clamp to the available screen
area, blob decoding, and both migrations against a scratch XDG_CONFIG_HOME.

+++ common/ttsettings.h
@@ -228,14 +228,14 @@ public:
   // carry no `Sec` suffix; preserved verbatim from gui/ttcutsettings.cpp
   // for round-trip compatibility with already-installed user settings.
   // No signals — none of the 12 fields have reactive UI dependents.
-  float   navBlackThreshold() const  { return mNavBlackThreshold; }
-  void    setNavBlackThreshold(float v);
+  double  navBlackThreshold() const  { return mNavBlackThreshold; }
+  void    setNavBlackThreshold(double v);
 
-  float   navSceneThreshold() const  { return mNavSceneThreshold; }
-  void    setNavSceneThreshold(float v);
+  double  navSceneThreshold() const  { return mNavSceneThreshold; }
+  void    setNavSceneThreshold(double v);
 
-  float   navLogoThreshold() const   { return mNavLogoThreshold; }
-  void    setNavLogoThreshold(float v);
+  double  navLogoThreshold() const   { return mNavLogoThreshold; }
+  void    setNavLogoThreshold(double v);
 
   bool    spDetectSilence() const      { return mSpDetectSilence; }
   void    setSpDetectSilence(bool v);
@@ -243,8 +243,8 @@ public:
   int     spSilenceThresholdDb() const { return mSpSilenceThresholdDb; }
   void    setSpSilenceThresholdDb(int v);
 
-  float   spSilenceMinDuration() const { return mSpSilenceMinDuration; }
-  void    setSpSilenceMinDuration(float v);
+  double  spSilenceMinDuration() const { return mSpSilenceMinDuration; }
+  void    setSpSilenceMinDuration(double v);
 
   bool    spDetectAudioChange() const  { return mSpDetectAudioChange; }
   void    setSpDetectAudioChange(bool v);
@@ -457,12 +457,12 @@ private:
   // (Tasks 4-5). The seven sp* fields share /Settings/StreamPoints. The two
   // extraFrame* fields share /Settings/Common (legacy keys without `Sec`
   // suffix — see public-section comment for detail).
-  float mNavBlackThreshold        = 0.980f;
-  float mNavSceneThreshold        = 0.300f;
-  float mNavLogoThreshold         = 0.500f;
+  double mNavBlackThreshold       = 0.980;
+  double mNavSceneThreshold       = 0.300;
+  double mNavLogoThreshold        = 0.500;
   bool  mSpDetectSilence          = true;
   int   mSpSilenceThresholdDb     = -75;
-  float mSpSilenceMinDuration     = 0.3f;
+  double mSpSilenceMinDuration    = 0.3;
   bool  mSpDetectAudioChange      = true;
   bool  mSpDetectAspectChange     = true;
   bool  mSpDetectPillarbox        = true;
